Fixed bug in mutation that missed cooperativity factors.
This was an interesting problem. When duplicating from a 2-site model with
co-op factor f12 = 10, it is clear that the "homologous" co-op factor in the
4-site model, f34, should also be 10. However, I was missing the fact that
f123,f124, and f1234, should be 10 as well, because the cooperativity between
sites 1 and 2 is conditionally independent of 3 and 4. The result of the bug
was that the duplicated curves were not steep enough - fixed it now, and
looks like it works well for the 4 to 8 site transition also.